Added file metadata check

+++ bot.py
@@ -11,6 +11,13 @@ from dotenv import load_dotenv, find_dotenv
 load_dotenv(find_dotenv())
 
 
+DONATION_LINK = 'https://google.com' # Paste your remote payment link here
+
+MAX_SIZE_MB = 50
+
+MAX_DURATION_SEC = 1800
+
+
 bot = telebot.TeleBot(os.getenv('TOKEN'))
 
 ydl_config = {
@@ -28,13 +35,10 @@ with open('responses.json', 'r') as f:
 	responses = json.load(f)
 
 
-MARKUP_DONATION = '💰 Donate'
-
-
 @bot.message_handler(commands=['start'])
 def send_welcome(message):
 	markup = types.ReplyKeyboardMarkup(resize_keyboard=True)
-	markup.add(types.KeyboardButton(MARKUP_DONATION))
+	markup.add(types.KeyboardButton('💰 Donate'))
 	bot.send_message(message.chat.id, random.choice(responses['greetings']), reply_markup=markup)
 
 
@@ -43,10 +47,10 @@ def send_help(message):
 	bot.send_message(message.chat.id, random.choice(responses['link']['entry']))
 
 
-@bot.message_handler(func=lambda m: m.text in [MARKUP_DONATION, '/donate'])
+@bot.message_handler(func=lambda m: m.text in ['💰 Donate', '/donate'])
 def handle_markup(message):
 	markup = types.InlineKeyboardMarkup()
-	markup.add(types.InlineKeyboardButton('Donation page', url=os.getenv('DONATION_LINK')))
+	markup.add(types.InlineKeyboardButton('Donation page', url=DONATION_LINK))
 	bot.send_message(message.chat.id, random.choice(responses['donation']['entry']), reply_markup=markup)
 
 
@@ -55,18 +59,28 @@ def handle_link(message):
 	temp_message = bot.send_message(message.chat.id, random.choice(responses['link']['waiting']))
 	with YoutubeDL(ydl_config) as ydl:
 		try:
+			ydl_info = ydl.extract_info(message.text, download=False)
+
+			ydl_audio_filesize = ydl_info.get('filesize') or ydl_info.get('filesize_approx')
+			if ydl_audio_filesize and ydl_audio_filesize > MAX_SIZE_MB * 1024 * 1024:
+				bot.delete_message(message.chat.id, temp_message.message_id)
+				bot.send_message(message.chat.id, random.choice(responses['link']['error']['filesize']))
+				return
+
+			ydl_duration = ydl_info.get('duration', 0)
+			if ydl_duration > MAX_DURATION_SEC:
+				bot.delete_message(message.chat.id, temp_message.message_id)
+				bot.send_message(message.chat.id, random.choice(responses['link']['error']['duration']))
+				return
+
 			ydl_info = ydl.extract_info(message.text, download=True)
+
 		except Exception:
 			bot.delete_message(message.chat.id, temp_message.message_id)
-			bot.send_message(message.chat.id, random.choice(responses['link']['error']))
+			bot.send_message(message.chat.id, random.choice(responses['link']['error']['link']))
 			return
-	"""ydl_dur = ydl_info['duration']
-	HH, mm, ss = ydl_dur//3600, (ydl_dur%3600)//60, ydl_dur%60
-	HHmmss_arr = [HH, mm, ss] if HH else [mm,ss]
-	ydl_output = ydl_info['title'] + ' (' + ':'.join(str(i).zfill(2) for i in HHmmss_arr) + ')'
-	bot.delete_message(message.chat.id, temp_message.message_id)
-	bot.send_message(message.chat.id, ydl_output)"""
-	ydl_path = os.path.join('temp', [i for i in os.listdir('temp') if i != '.gitkeep'][0])
+
+	ydl_path = os.path.join('temp', f'{ydl_info['title']}.mp3')
 	with open(ydl_path, 'rb') as f:
 		bot.send_audio(message.chat.id, f)
 	bot.delete_message(message.chat.id, temp_message.message_id)
